Move `Input` to swc_common
