+++ common/src/lib.rs
@@ -34,6 +34,7 @@ pub mod comments;
 pub mod errors;
 #[cfg(feature = "fold")]
 pub mod fold;
+pub mod input;
 pub mod macros;
 mod pos;
 mod rustc_data_structures;
